+++ src/Domain/TeamMatch/RubberDiscipline.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DimiPress\Rally\Domain\TeamMatch;
+
+enum RubberDiscipline: string
+{
+    case Singles = 'singles';
+    case Doubles = 'doubles';
+}