+++ src/Domain/TeamMatch/TeamMatchSide.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DimiPress\Rally\Domain\TeamMatch;
+
+enum TeamMatchSide: string
+{
+    case Home = 'home';
+    case Away = 'away';
+}